reporter: Prepare for implementing alternative notice reporters
Move the logic that creates the notice report to a new `NoticeProcessor`
class and make the notice reporter class abstract. This allows to easily
add subclasses of `AbstractNoticeReporter` which provide their own
`NoticeProcessor`s.
